fix(cli): collect server node assets
