+++ packages/server/src/node-assets.ts
@@ -1,0 +1,5 @@
+import { getCoreNodeAssets } from "@opencode-ai/core/node-assets"
+
+export function getServerNodeAssets() {
+  return getCoreNodeAssets()
+}
